[Task 4] Fix review
